MAINT: Move pr_reminder.py into slack_reminder_api
Moving this file into a better named directory
