+++ cloud-chatops/src/events.py
@@ -9,7 +9,7 @@ from slack_sdk import WebClient
 import schedule
 
 from data import User
-from features.pr_reminder import PRReminder
+from slack_reminder_api.pr_reminder import PRReminder
 from find_prs import FindPRs
 from read_data import get_config, get_token
